move components away from routes folder
