+++ app/components/Program/AttendanceChartBars.tsx
